# spec/openapi.yaml
openapi: "3.0.3"
info:
  title: LANSheet Local Node API
  version: "0.1.0"
servers:
  - url: "http://localhost:8443"
paths:
  /v1/documents:
    post:
      summary: create document metadata
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentCreate"
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentSummary"
    get:
      summary: list documents
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentSummary"

  /v1/documents/{doc_id}:
    get:
      summary: get document metadata and latest block info
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /v1/documents/{doc_id}/blocks:
    get:
      summary: fetch blocks since a known hash
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: string
      responses:
        "200":
          description: blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockJSON"

    post:
      summary: submit a block
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockJSON"
      responses:
        "201":
          description: accepted

  /v1/documents/{doc_id}/export/xlsx:
    get:
      summary: export workbook as XLSX
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: xlsx binary
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary

  /v1/nodes/peers:
    get:
      summary: list peers
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Peer"

components:
  schemas:
    DocumentCreate:
      type: object
      properties:
        name:
          type: string
        owner:
          type: string

    DocumentSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        owner:
          type: string
        head_block:
          type: string

    Document:
      allOf:
        - $ref: "#/components/schemas/DocumentSummary"
        - type: object
          properties:
            created_at:
              type: string
            sheets:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  rows:
                    type: integer
                  cols:
                    type: integer

    BlockJSON:
      type: object
      properties:
        header:
          type: object
          properties:
            doc_id:
              type: string
            prev_hash:
              type: string
            author:
              type: string
            author_pubkey:
              type: string
            timestamp:
              type: string
            version:
              type: integer
        payload:
          type: object
          properties:
            crdt_update_b64:
              type: string
            ops:
              type: array
              items:
                type: object
        merkle_root:
          type: string
        signature:
          type: string

    Peer:
      type: object
      properties:
        id:
          type: string
        addr:
          type: string
        last_seen:
          type: string